Fix type error

+++ src/android/LocationManager.java
@@ -85,8 +85,8 @@ public class LocationManager extends CordovaPlugin implements BeaconConsumer {
 
     private static final String BACKGROUND_BETWEEN_SCAN_PERIOD_NAME = "com.unarin.cordova.beacon.android.altbeacon.BackgroundBetweenScanPeriod";
     private static final String BACKGROUND_SCAN_PERIOD_NAME = "com.unarin.cordova.beacon.android.altbeacon.BackgroundScanPeriod";
-    public static final long DEFAULT_BACKGROUND_BETWEEN_SCAN_PERIOD = 300000;
-    public static final long DEFAULT_BACKGROUND_SCAN_PERIOD = 10000;
+    public static final int DEFAULT_BACKGROUND_BETWEEN_SCAN_PERIOD = 300000;
+    public static final int DEFAULT_BACKGROUND_SCAN_PERIOD = 10000;
 
     private static int CDV_LOCATION_MANAGER_DOM_DELEGATE_TIMEOUT = 30;
     private static final int BUILD_VERSION_CODES_M = 23;
